requirements.txt
blinker==1.6.2
click==8.1.6
distro==1.8.0
Flask==2.3.2
gunicorn==21.2.0
icalendar==5.0.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
packaging==23.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabula-py==2.7.0
tzdata==2023.3
Werkzeug==2.3.6
